refactor: Refactor annotation code to make it less monolithic.
This doesn't change any exposed methods.
